Fix incorrect config filename in update script
The update script backs up the config file to `/root/config.bak`, but
then tries to restore it from `/root/config.yml.bak` (which obviously
won't be found).

This looks like a typo in the backup command, so update that command to
use the same filename as the restore command.

Signed-off-by: Ross Simpson <ross@rosssimpson.com>